Use healthchecks to work around ssh key gen race conditions

+++ frontend/docker-entrypoint.sh
@@ -8,11 +8,14 @@ sudo ssh-keygen -A
 sudo service ssh start
 
 # Generate SSH key and set up shared authorized_keys
-mkdir -p /home/admin/.ssh
-ssh-keygen -t rsa -f /home/admin/.ssh/id_rsa -N ""
-cp /home/admin/.ssh/id_rsa.pub /home/admin/.ssh/authorized_keys
-chown -R admin:admin /home/admin/.ssh
-chmod 700 /home/admin/.ssh
-chmod 600 /home/admin/.ssh/authorized_keys
+if [ ! -f /home/admin/.ssh/id_rsa ]; then
+    ssh-keygen -t rsa -f /home/admin/.ssh/id_rsa -N ""
+    cp /home/admin/.ssh/id_rsa.pub /home/admin/.ssh/authorized_keys
+    chown -R admin:admin /home/admin/.ssh
+    chmod 700 /home/admin/.ssh
+    chmod 600 /home/admin/.ssh/authorized_keys
+fi
+
+touch /tmp/ssh-ready
 
 tail -f /dev/null
